Initial bofire_mixed refactor
