+++ pyproject.toml
@@ -21,7 +21,8 @@ dependencies = [
     "bofire[cheminfo,optimization]<1.0.0,>=0.0.14",
     "numba-scipy<1.0.0,>=0.4.0",
     "entmoot<3.0.0,>=2.0.2",
-    "bofire-mixed-baselines",
+    "arviz>=0.21.0",
+    "pymc-bart>=0.9.2",
 ]
 name = "bark"
 version = "0.1.0"
@@ -38,20 +39,15 @@ ignore-init-module-imports = true
 "__init__.py" = ["F401"]
 
 
-[tool.pdm.dev-dependencies]
-dev = [
-    "pytest<9.0.0,>=8.0.0",
-    "ruff<1.0.0,>=0.2.0",
-    "pre-commit<4.0.0,>=3.6.0",
-    "ipython<9.0.0,>=8.21.0",
-    "pytest-benchmark<5.0.0,>=4.0.0",
-]
-
-[tool.uv.sources]
-bofire-mixed-baselines = { git = "https://github.com/TobyBoyne/bofire-mixed-baselines.git" }
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+# Defines the targets for the wheel build.
+include = ["src/bark/**", "src/bofire_mixed/**"]
+sources = ["src"]
+packages = ["src/bark", "src/bofire_mixed"]
 
 
 [dependency-groups]
