+++ src/bofire_mixed/benchmarks/MAX_bandit.py
