+++ src/bofire_mixed/benchmarks/data/MAX_data.json
