+++ src/bofire_mixed/benchmarks/xgboost_mnist.py
@@ -12,7 +12,7 @@ from bofire.data_models.features.api import (
 from bofire.data_models.objectives.api import MinimizeObjective
 from sklearn import datasets, metrics, model_selection
 
-from bark.bofire_utils.domain import build_integer_input
+from bofire_mixed.domain import build_integer_input
 
 
 class XGBoostMNIST(Benchmark):
