+++ tests/optimization/test_optimality.py
@@ -4,12 +4,12 @@ from bofire.benchmarks.api import Benchmark, Himmelblau
 from bofire.data_models.domain.api import Domain
 
 from bark.benchmarks import StyblinskiTang
-from bark.bofire_utils.domain import get_feature_types_array
 from bark.fitting.bark_sampler import BARKTrainParams, run_bark_sampler
 from bark.forest import batched_forest_gram_matrix, create_empty_forest
 from bark.optimizer import propose
 from bark.optimizer.opt_core import get_opt_core_from_domain
 from bark.optimizer.opt_model import build_opt_model_from_forest
+from bofire_mixed.domain import get_feature_types_array
 
 
 def train_data(benchmark: Benchmark, n: int) -> tuple[np.ndarray, np.ndarray]:
